Explain middleware `Use()` required ordering (#131)
This addition explains how `Echo#Use()` must be called in order for the middleware to be executed. Previous issues (https://github.com/labstack/echo/issues/1304) have referenced this and the explanation has been to look at the unit tests. Having the behavior documented will help users understand how to correctly use this part of the library.